Prune trivial types which aren't likely to be used.
