+++ src/genjax/_src/core/generative/__init__.py
@@ -27,25 +27,25 @@ from .core import (
     Argdiffs,
     Arguments,
     Constraint,
-    EditRequest,
     EmptySample,
-    IncrementalRegenerateRequest,
-    IncrementalUpdateRequest,
     MaskedSample,
     Projection,
     R,
     Retdiff,
     Sample,
     Score,
-    UnhandledEditRequestException,
     Weight,
 )
 from .functional_types import Mask
 from .generative_function import (
+    EditRequest,
     GenerativeFunction,
     GenerativeFunctionClosure,
     IgnoreKwargs,
+    IncrementalRegenerateRequest,
+    IncrementalUpdateRequest,
     Trace,
+    UnhandledEditRequestException,
 )
 
 __all__ = [
