+++ src/genjax/_src/core/generative/generative_function.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 from abc import abstractmethod
+from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
 import jax
@@ -27,9 +28,6 @@ from genjax._src.core.generative.core import (
     Argdiffs,
     Arguments,
     Constraint,
-    EditRequest,
-    IncrementalRegenerateRequest,
-    IncrementalUpdateRequest,
     Projection,
     Retdiff,
     Sample,
@@ -46,6 +44,7 @@ from genjax._src.core.typing import (
     InAxes,
     Int,
     PRNGKey,
+    Self,
     String,
     TypeVar,
 )
@@ -156,9 +155,9 @@ class Trace(Generic[R], Pytree):
     def edit(
         self,
         key: PRNGKey,
-        request: EditRequest,
+        request: "EditRequest",
         argdiffs: Argdiffs,
-    ) -> tuple["Trace[R]", Weight, Retdiff[R], EditRequest]:
+    ) -> tuple[Self, Weight, Retdiff[R], "EditRequest"]:
         """
         This method calls out to the underlying [`GenerativeFunction.edit`][genjax.core.GenerativeFunction.edit] method - see [`EditRequest`][genjax.core.EditRequest] and [`edit`][genjax.core.GenerativeFunction.edit] for more information.
         """
@@ -174,7 +173,7 @@ class Trace(Generic[R], Pytree):
         key: PRNGKey,
         constraint: ChoiceMap,
         argdiffs: tuple[Any, ...] | None = None,
-    ) -> tuple["Trace[R]", Weight, Retdiff[R], ChoiceMap]:
+    ) -> tuple[Self, Weight, Retdiff[R], ChoiceMap]:
         """
         This method calls out to the underlying [`GenerativeFunction.edit`][genjax.core.GenerativeFunction.edit] method - see [`EditRequest`][genjax.core.EditRequest] and [`edit`][genjax.core.GenerativeFunction.edit] for more information.
         """
@@ -183,14 +182,14 @@ class Trace(Generic[R], Pytree):
             self,
             constraint,
             Diff.tree_diff_no_change(self.get_args()) if argdiffs is None else argdiffs,
-        )
+        )  # pyright: ignore[reportReturnType]
 
     def regenerate(
-        self,
+        self: Self,
         key: PRNGKey,
         projection: Selection,
         argdiffs: tuple[Any, ...] | None = None,
-    ) -> tuple["Trace[R]", Weight, Retdiff[R], ChoiceMap]:
+    ) -> tuple[Self, Weight, Retdiff[R], ChoiceMap]:
         """
         This method calls out to the underlying [`GenerativeFunction.edit`][genjax.core.GenerativeFunction.edit] method - see [`EditRequest`][genjax.core.EditRequest] and [`edit`][genjax.core.GenerativeFunction.edit] for more information.
         """
@@ -199,7 +198,7 @@ class Trace(Generic[R], Pytree):
             self,
             projection,
             Diff.tree_diff_no_change(self.get_args()) if argdiffs is None else argdiffs,
-        )
+        )  # pyright: ignore[reportReturnType]
 
     def project(
         self,
@@ -227,6 +226,62 @@ class Trace(Generic[R], Pytree):
     @property
     def batch_shape(self):
         return len(self.get_score())
+
+
+#################
+# Edit requests #
+#################
+
+
+class EditRequest(Pytree):
+    """
+    An `EditRequest` is a request to edit a trace of a generative function. Generative functions respond to instances of subtypes of `EditRequest` by providing an [`edit`][genjax.core.GenerativeFunction.edit] implementation.
+
+    Updating a trace is a common operation in inference processes, but naively mutating the trace will invalidate the mathematical invariants that Gen retains. `EditRequest` instances denote requests for _SMC moves_ in the framework of [SMCP3](https://proceedings.mlr.press/v206/lew23a.html), which preserve these invariants.
+    """
+
+    @abstractmethod
+    def edit(
+        self,
+        key: PRNGKey,
+        trace: Trace[R],
+        argdiffs: Argdiffs,
+    ) -> tuple[Trace[R], Weight, Retdiff[R], "EditRequest"]:
+        pass
+
+
+@Pytree.dataclass(match_args=True)
+class IncrementalUpdateRequest(EditRequest):
+    constraint: Constraint
+
+    def edit(
+        self,
+        key: PRNGKey,
+        trace: Trace[R],
+        argdiffs: Argdiffs,
+    ) -> tuple[Trace[R], Weight, Retdiff[R], "EditRequest"]:
+        gen_fn = trace.get_gen_fn()
+        return gen_fn.edit(key, trace, self, argdiffs)
+
+
+@Pytree.dataclass(match_args=True)
+class IncrementalRegenerateRequest(EditRequest):
+    projection: Projection[Any]
+
+    def edit(
+        self,
+        key: PRNGKey,
+        trace: Trace[R],
+        argdiffs: Argdiffs,
+    ) -> tuple[Trace[R], Weight, Retdiff[R], "EditRequest"]:
+        gen_fn = trace.get_gen_fn()
+        return gen_fn.edit(key, trace, self, argdiffs)
+
+
+@dataclass
+class UnhandledEditRequestException(Exception):
+    gen_fn: "GenerativeFunction[Any]"
+    edit_request: EditRequest
 
 
 #######################
